sysutils/py-puremagic: import py-puremagic-1.15
Packaged for wip by Olaf Seibert and myself.

puremagic is a pure python module that will identify a file based off its
magic numbers.

It is designed to be minimalistic and inherently cross platform
compatible. It is also designed to be a stand in for python-magic, it
incorporates the functions from_file(filename[, mime]) and
from_string(string[, mime]) however the magic_file() and magic_string()
are more powerful and will also display confidence and duplicate
matches.

It does NOT try to match files off non-magic string. In other words it
will not search for a string within a certain window of bytes like
others might.