py-openid: update to 3.2.0
3.0.9
Compatibility fix for `HTMLParser`

3.0.6
Improvements to stores, tests, and fetchers

3.0.3
Improvements to admin test runner

3.0.2
All the tests that require no more than the standard library pass -- however
modules that depend on external packages (such as the pyCURL fetcher or the
MySQL / PostgreSQL stores) haven't been updated and may not work.

3.0.1
This implementation of OpenID has been ported to Python 3 -- all but one test
is known to pass.